+++ externallib.php
@@ -31,6 +31,7 @@ require_once($CFG->dirroot . '/local/booking/lib.php');
 use local_booking\external\logentry_exporter;
 use local_booking\local\logbook\forms\create as update_logentry_form;
 use local_booking\local\logbook\entities\logbook;
+use local_booking\local\participant\entities\participant;
 use local_booking\local\participant\entities\student;
 use local_booking\local\subscriber\entities\subscriber;
 
@@ -289,15 +290,23 @@ class local_booking_external extends external_api {
         $result = true;
         $warnings = array();
         $errorcode = '';
+
+        // get PIREP integrated info
         $logentry = (new logbook($courseid, $userid))->create_logentry();
-        $rec = subscriber::get_integrated_data('pireps', 'pirep', $pirep);
-        if (!empty($rec)) {
-            $logentry->read($rec);
-            // get engine type integrated data
-            if (subscriber::has_integration('pilot')) {
-                $pilotrec = subscriber::get_integrated_data('pilot', 'pilot', $logentry->pilot_id);
+        $pireprec = subscriber::get_integrated_data('pireps', 'pirepinfo', $pirep);
+
+        if (!empty($pireprec)) {
+
+            // get logentry data from the PIREP record
+            $logentry->read($pireprec);
+
+            // get pilot integrated info
+            if (subscriber::has_integration('pilots')) {
+                $pilotrec = subscriber::get_integrated_data('pilots', 'pilotinfo', $logentry->pilot_id);
                 $alternatename = $pilotrec['alternatename'];
+
                 if (core_user::get_user($userid, 'alternatename')->alternatename == $alternatename) {
+
                     // get engine type integrated data
                     if (subscriber::has_integration('aircraft')) {
                         $enginetyperec = subscriber::get_integrated_data('aircraft', 'aircraftinfo', $logentry->get_aircraft());
@@ -310,6 +319,7 @@ class local_booking_external extends external_api {
                     $data['view'] = 'summary';
                     $data['nullable'] = false;
                     list($data, $template) = get_logentry_view($courseid, $userid, $data);
+
                 } else {
                     $result = false;
                     $errorcode = 'errorp1pirepwrongpilot';
@@ -805,6 +815,75 @@ class local_booking_external extends external_api {
      * @since Moodle 2.5
      */
     public static function update_user_group_returns() {
+        return new external_single_structure(
+            array(
+                'result' => new external_value(PARAM_BOOL, get_string('processingresult', 'local_booking')),
+                'warnings' => new external_warnings()
+            )
+        );
+    }
+
+    /**
+     * Returns description of method parameters.
+     *
+     * @return external_function_parameters
+     */
+    public static function update_profile_comment_parameters() {
+        return new external_function_parameters(array(
+                'courseid' => new external_value(PARAM_INT, 'The course id', VALUE_DEFAULT),
+                'userid' => new external_value(PARAM_INT, 'The student id', VALUE_DEFAULT),
+                'comment' => new external_value(PARAM_RAW, 'The comment text', VALUE_DEFAULT),
+            )
+        );
+    }
+
+    /**
+     * Update user group membership add/remove for the course.
+     *
+     * @param int    $courseid   The course id.
+     * @param int    $userid     The user id.
+     * @param string $comment    The comment text.
+     * @return bool $result      The comment save was successful.
+     * @throws moodle_exception if user doesnt have the permission to create events.
+     */
+    public static function update_profile_comment(int $courseid, int $userid, string $comment) {
+        global $COURSE, $PAGE;
+
+        // Parameter validation.
+        $params = self::validate_parameters(self::update_profile_comment_parameters(), array(
+            'courseid'=> $courseid,
+            'userid'  => $userid,
+            'comment' => $comment,
+            )
+        );
+
+        $warnings = array();
+
+        $context = context_course::instance($courseid);
+        self::validate_context($context);
+        $PAGE->set_url('/local/booking/');
+
+        // define subscriber globally
+        if (empty($COURSE->subscriber))
+            $COURSE->subscriber = new subscriber($courseid);
+
+        // add/remove student to group
+        $participant = new participant($COURSE->subscriber, $userid);
+        $result = $participant->update_comment($comment);
+
+        return array(
+            'result' => $result,
+            'warnings' => $warnings
+        );
+    }
+
+    /**
+     * Returns description of method result value.
+     *
+     * @return external_description.
+     * @since Moodle 2.5
+     */
+    public static function update_profile_comment_returns() {
         return new external_single_structure(
             array(
                 'result' => new external_value(PARAM_BOOL, get_string('processingresult', 'local_booking')),