+++ CHANGELOG.md
@@ -6,6 +6,18 @@
 
 - cron task background processing of session booking scheduled notifications (availability postings)
 
+## [2022102600] - UNRELEASED 2022-10-26
+
+### Added
+
+- Students progression page, a read-only view for students progression
+- ability to save comments regarding a student in his/her profile (i.e. temporary leave)
+
+### Changed
+
+- replaced ATO name with the plugin name as the section (category) label in the course settings page
+- removed ATO plugin settings information from config.json
+
 ## [2022102100] - RELEASED 2022-10-21
 
 ### Fixed